// client/src/types/authTypes.ts
export interface LoginTypes {
  email: string;
  password: string;
}

export interface LoginInputTypes {
  name: string;
  placeholder: string;
  type: string;
}

export interface SignupTypes {
  username: string;
  email: string;
  password: string;
  confirmpassword?: string;
}

export interface SignupInputTypes {
  name: string;
  placeholder: string;
  type: string;
}